Spring boot version updated, Spring oauth2 auth server version updated. Added client logout method
