--- src/main/java/pl/achrzanowski/moneymanagementauthorizationservice/config/DefaultSecurityConfig.java
@@ -1,41 +1,0 @@
-package pl.achrzanowski.moneymanagementauthorizationservice.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.provisioning.JdbcUserDetailsManager;
-import org.springframework.security.provisioning.UserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-
-import javax.sql.DataSource;
-
-@EnableWebSecurity
-public class DefaultSecurityConfig {
-
-    @Autowired
-    private DataSource dataSource;
-
-    @Bean
-    UserDetailsManager userDetailsManager(){
-        return new JdbcUserDetailsManager(dataSource);
-    }
-
-    @Bean
-    SecurityFilterChain defaultSecurityFilterChain(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity
-                .authorizeRequests(
-                        authorizeRequests -> authorizeRequests
-                                .mvcMatchers("/content/**").permitAll()
-                                .anyRequest().authenticated())
-                .csrf().disable()
-                .headers().frameOptions().disable()
-                .and()
-                .formLogin(form -> form
-                                .loginPage("/login")
-                                .permitAll());
-        return httpSecurity.build();
-    }
-
-
-}
